Module gate::asset_id

Traits related to asset enums that are generated by "gate_build".

The user is not expected to implement these traits themselves directly. Rather, the generated code from "gate_build" will implement these traits. Gate build will generate four enums: SpriteId, MusicId, and SoundId, that implement IdU16. It will also generate a collection type AssetId that implements AppAssetId.

Traits

AppAssetId
A convenience collection of all asset ID types.

IdU16
A value that can be converted to a u16 ID, implemented by asset enums.